pyproject.toml
[project]
name = "apps-service"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "django==5.1.1",
    "django-ninja==1.3.0",
    "bigcommerce==0.23.4",
    "ShopifyAPI==12.6.0",
]
readme = "README.md"
requires-python = "==3.12.0"



[tool.rye]
managed = true
virtualenv = true
dev-dependencies = []

[tool.setuptools]
py-modules = []